copy/telegram-bot/requirements.txt
# Python Telegram Bot Dependencies

# Main telegram bot library
python-telegram-bot==20.7

# Environment variable management
python-dotenv==1.0.0

# Optional: For more advanced features
# aiohttp==3.9.1  # Async HTTP client
# requests==2.31.0  # HTTP library
